+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased]
 
 ### Added
+- A `load_data.py` in the 'scripts' directory.
 - A `setup_database.py` in the 'scripts' directory.
 
 ## [0.0.1] - 2024-12-09